# docs/agent_os_openapi.yaml
openapi: 3.1.0
info:
  title: Enstack Agent OS API
  version: 1.0.0-draft.1
  summary: Give your AI agents real money — and real control.
  description: |
    # Enstack Agent OS — /v1 API

    The Agent OS API lets a verified **Owner** (a person or business) provision
    **Agents** — AI sub-principals that hold a Balance, spend on a Visa card,
    receive into a virtual bank account, and move money over local and
    cross-border rails via **Enstack Routing** — all inside an Owner-defined
    **Policy** (limits, allowlists, human approvals, full audit).

    ## Identity model
    - **Owner** — the KYC/KYB-verified principal. Owns all funds and all liability.
    - **Agent** — a sub-principal under an Owner. No separate KYC. Every agent
      action is attributable to both the agent and its owner.
    - **Credential** — a scoped API key or MCP token bound to exactly one agent
      (or to the owner). Credentials carry scopes; scopes are further constrained
      by the agent's Policy at execution time.
    - **Policy** — the authorization layer: per-transaction / daily / monthly
      limits, counterparty and category allowlists, and human-in-the-loop
      approval rules. Policy is versioned; every evaluation is audited.

    ## Conventions
    - **Versioning**: URL prefix `/v1/...`. Breaking changes ship as `/v2`.
    - **Idempotency**: `Idempotency-Key` header is **required on every mutating
      request** (POST/PUT/PATCH). Window: 24h, keyed by `(credential, key)`.
...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /agents/{agent_id}/credentials:
    parameters:
      - $ref: '#/components/parameters/agentId'
    post:
      operationId: createCredential
      tags: [Credentials]
      summary: Issue a credential for an agent
      description: |
        Owner scope only. Returns the secret **once**. `kind: mcp` credentials
        authenticate against the Agent OS MCP server; `kind: api` against this
        REST surface. Scopes requested here are intersected with the agent's
        Policy at every call.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CredentialCreate' }
      responses:
        '201':
          description: Credential created. `secret` is returned only in this response.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CredentialWithSecret' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listCredentials
      tags: [Credentials]
      summary: List an agent's credentials
      responses:
        '200':
          description: Credentials (secrets never included).
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Credential' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /credentials/{credential_id}/revoke:
    post:
      operationId: revokeCredential
      tags: [Credentials]
      summary: Revoke a credential
      description: Owner scope only. Irreversible; issue a new credential to restore access.
      parameters:
        - name: credential_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Revoked credential.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Credential' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Policy ─────────────────────────────
...
  /payments:
    post:
      operationId: createPayment
      tags: [Payments]
      summary: Create a payment
      description: |
        Sends funds from an agent's balance to a counterparty over the best
        available rail (local instant rails, bank transfer, or cross-border via
        Enstack Routing — rail selection is automatic unless pinned).

        **Policy evaluation happens here.** Outcomes:
        - Within policy → `201`, status `processing`.
        - Exceeds a limit or misses the allowlist but is approvable →
          `202`, status `pending_approval`; an Approval object is created and
          the owner is notified.
        - Hard-denied by policy → `403` with code `policy_denied`.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentCreate' }
      responses:
        '201':
          description: Payment accepted and processing.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        '202':
          description: Payment created but held pending owner approval.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listPayments
      tags: [Payments]
      summary: List payments
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/starting_after'
        - name: agent_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PaymentStatus' }
        - name: created_after
          in: query
          schema: { type: string, format: date-time }
        - name: created_before
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Payment list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Payment' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /payments/{payment_id}:
    get:
      operationId: getPayment
      tags: [Payments]
      summary: Retrieve a payment
      parameters:
        - name: payment_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The payment, including its routing and status timeline.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        default: { $ref: '#/components/responses/Error' }
  /payments/{payment_id}/cancel:
    post:
      operationId: cancelPayment
      tags: [Payments]
      summary: Cancel a payment
      description: Only payments in `pending_approval` or pre-dispatch `processing` can be cancelled. Cancellation of a dispatched payment returns `409`.
      parameters:
        - name: payment_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Cancelled payment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
        default: { $ref: '#/components/responses/Error' }

  # ─────────────────────────── Counterparties ───────────────────────────
  /counterparties:
    post:
      operationId: createCounterparty
      tags: [Counterparties]
      summary: Create a counterparty
      description: |
        Saves a payee (bank account, wallet, or card payout destination).
        Counterparties created by an agent credential start `unverified`;
        Policy can require `verified` (owner-confirmed) counterparties only.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CounterpartyCreate' }
      responses:
        '201':
          description: Counterparty created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Counterparty' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listCounterparties
      tags: [Counterparties]
      summary: List counterparties
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/starting_after'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/CounterpartyStatus' }
      responses:
        '200':
          description: Counterparty list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Counterparty' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /counterparties/{counterparty_id}:
    get:
      operationId: getCounterparty
      tags: [Counterparties]
      summary: Retrieve a counterparty
      parameters:
        - name: counterparty_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The counterparty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Counterparty' }
        default: { $ref: '#/components/responses/Error' }
  /counterparties/{counterparty_id}/verify:
    post:
      operationId: verifyCounterparty
      tags: [Counterparties]
      summary: Verify a counterparty
      description: Owner scope only. Marks the counterparty owner-confirmed, making it eligible under `verified_counterparties_only` policies.
      parameters:
        - name: counterparty_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Verified counterparty.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Counterparty' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Cards ─────────────────────────────
  /agents/{agent_id}/cards:
    parameters:
      - $ref: '#/components/parameters/agentId'
    post:
      operationId: createCard
      tags: [Cards]
      summary: Issue a card for an agent
      description: |
        Issues a virtual Visa card funded by the agent's balance. Card
        authorizations are evaluated against the agent's Policy in real time
        (limits, merchant-category allowlists). Physical cards: Phase 2.
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CardCreate' }
      responses:
        '201':
          description: Card issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Card' }
        default: { $ref: '#/components/responses/Error' }
    get:
      operationId: listCards
      tags: [Cards]
      summary: List an agent's cards
      responses:
        '200':
          description: Card list.
          content:
            application/json:
              schema:
                type: object
                required: [data, has_more]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Card' }
                  has_more: { type: boolean }
                  next_cursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
  /cards/{card_id}:
    get:
      operationId: getCard
      tags: [Cards]
      summary: Retrieve a card
      description: PAN and CVV are never returned by this endpoint; sensitive card details are revealed only through the secure display flow in the app or a PCI-scoped reveal token.
      parameters:
        - name: card_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The card (masked).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Card' }
        default: { $ref: '#/components/responses/Error' }
  /cards/{card_id}/freeze:
    post:
      operationId: freezeCard
      tags: [Cards]
      summary: Freeze a card
      description: Available to both owner and the card's own agent (an agent may freeze, but not unfreeze, its own card).
      parameters:
        - name: card_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Frozen card.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Card' }
        default: { $ref: '#/components/responses/Error' }
  /cards/{card_id}/unfreeze:
    post:
      operationId: unfreezeCard
      tags: [Cards]
      summary: Unfreeze a card
      description: Owner scope only.
      parameters:
        - name: card_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '200':
          description: Active card.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Card' }
        default: { $ref: '#/components/responses/Error' }

  # ──────────────────────── Virtual Accounts ────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        Owner key (`ok_live_…`) or agent credential (`ac_live_…`).
        Agent credentials are policy-bound: server-side Policy evaluation
        applies to every call regardless of credential scopes.
  parameters:
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key (UUIDv4 recommended). Required on every
        mutating request. 24h dedup window keyed by (credential, key).
        Same key + same body replays the original response; same key +
        different body returns 409.
      schema:
        type: string
        minLength: 16
        maxLength: 128
    agentId:
      name: agent_id
      in: path
      required: true
      schema: { type: string }
      description: Agent identifier, e.g. `agt_01J….`
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    starting_after:
      name: starting_after
      in: query
      schema: { type: string }
      description: Cursor — the `next_cursor` value from the previous page.
  responses:
    Error:
      description: |
        Error envelope. Notable codes: `policy_denied` (403, hard policy
        block), `insufficient_funds` (409), `idempotency_conflict` (409,
        key reuse with different body), `quote_expired` (409),
        `approval_required` is NOT an error — held actions return 202.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples: [policy_denied, insufficient_funds, quote_expired, idempotency_conflict, not_found, unauthorized, validation_error, rate_limited]
        message:
          type: string
          description: Human-readable description.
        details:
          type: [object, 'null']
          description: Field-level or context detail (e.g. which policy rule denied).
          additionalProperties: true
        request_id:
          type: string
          description: Correlate with support and the activity feed.
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: string
          pattern: '^-?\d+(\.\d+)?$'
          description: Decimal string. Never a float.
          examples: ["1500.00", "0.4275"]
        currency:
          type: string
          description: ISO 4217 code.
          examples: [USD, PHP, HKD, EUR, SGD]
...
    CredentialKind:
      type: string
      enum: [api, mcp]
      description: "`api` authenticates against this REST surface; `mcp` against the Agent OS MCP server."
    CredentialScope:
      type: string
      enum:
        - balance:read
        - quotes:create
        - payments:create
        - payments:read
        - transfers:create
        - transfers:read
        - counterparties:create
        - counterparties:read
        - cards:read
        - cards:freeze
        - approvals:read
        - activity:read
        - policy:read
      description: Requested capability. Effective capability = scopes ∩ Policy.
    Credential:
      type: object
      required: [id, agent_id, kind, scopes, status, created_at]
      properties:
        id: { type: string, examples: [crd_01JABCDXYZ] }
        agent_id: { type: string }
        kind: { $ref: '#/components/schemas/CredentialKind' }
        label: { type: [string, 'null'], description: Human label, e.g. "prod procurement agent". }
        scopes:
          type: array
          items: { $ref: '#/components/schemas/CredentialScope' }
        status: { type: string, enum: [active, revoked] }
        last_used_at: { type: [string, 'null'], format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
        created_at: { type: string, format: date-time }
    CredentialCreate:
      type: object
      required: [kind, scopes]
      properties:
        kind: { $ref: '#/components/schemas/CredentialKind' }
        label: { type: string, maxLength: 120 }
        scopes:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CredentialScope' }
        expires_at:
          type: string
          format: date-time
          description: Optional hard expiry. Recommended for MCP tokens.
    CredentialWithSecret:
      allOf:
        - $ref: '#/components/schemas/Credential'
        - type: object
          required: [secret]
          properties:
            secret:
              type: string
              description: The bearer token. Shown once; store securely.
              examples: [ac_live_9f3K…]
...
    PaymentStatus:
      type: string
      enum: [pending_approval, processing, completed, failed, cancelled, returned]
    PaymentCreate:
      type: object
      required: [agent_id, counterparty_id]
      properties:
        agent_id: { type: string }
        counterparty_id: { type: string }
        quote_id:
          type: string
          description: Executes at the quoted rate. Required for cross-currency payments; omit for same-currency.
        amount:
          $ref: '#/components/schemas/Money'
          description: Required when no quote_id is given (same-currency payments).
        reference:
          type: string
          maxLength: 140
          description: Appears on the recipient statement where the rail supports it.
        purpose:
          type: string
          maxLength: 280
          description: Free-text purpose for the audit trail and the approval card the owner sees.
    Payment:
      type: object
      required: [id, agent_id, counterparty_id, source_amount, status, created_at]
      properties:
        id: { type: string, examples: [pay_01JABCDXYZ] }
        agent_id: { type: string }
        credential_id:
          type: string
          description: The credential that initiated the payment. Attribution is never optional.
        counterparty_id: { type: string }
        quote_id: { type: [string, 'null'] }
        source_amount: { $ref: '#/components/schemas/Money' }
        destination_amount: { $ref: '#/components/schemas/Money' }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        approval_id:
          type: [string, 'null']
          description: Present when the payment was (or is) held for approval.
        failure_reason: { type: [string, 'null'] }
        rail:
          type: [string, 'null']
          description: The rail Enstack Routing selected, e.g. `fps`, `pix`, `upi`, `sepa_instant`, `ach`, `bank_transfer`, `ewallet`.
        reference: { type: [string, 'null'] }
        purpose: { type: [string, 'null'] }
        timeline:
          type: array
          description: Ordered status transitions with timestamps — the tracking view.
          items:
            type: object
            required: [status, at]
            properties:
              status: { type: string }
              at: { type: string, format: date-time }
              detail: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
    CounterpartyStatus:
      type: string
      enum: [unverified, verified, blocked]
    CounterpartyCreate:
      type: object
      required: [display_name, destination]
      properties:
        display_name: { type: string, maxLength: 120 }
        destination:
          type: object
          required: [type]
          properties:
            type:
              type: string
              enum: [bank_account, ewallet, card_payout]
            currency: { type: string }
            country: { type: string }
            account_name: { type: string }
            account_number: { type: string }
            bank_identifier: { type: string, description: SWIFT/BIC or local bank code. }
            ewallet_provider: { type: string }
            ewallet_id: { type: string }
        metadata:
          type: object
          additionalProperties: { type: string }
    Counterparty:
      type: object
      required: [id, display_name, destination_summary, status, created_at]
      properties:
        id: { type: string, examples: [cpt_01JABCDXYZ] }
        display_name: { type: string }
        destination_summary:
          type: object
          description: Masked destination details (never full account numbers on read).
          properties:
            type: { type: string }
            currency: { type: string }
            country: { type: string }
            masked_identifier: { type: string, examples: ["••••4821"] }
        status: { $ref: '#/components/schemas/CounterpartyStatus' }
        created_by:
          type: object
          properties:
            actor_type: { type: string, enum: [owner, agent] }
            actor_id: { type: string }
        created_at: { type: string, format: date-time }
    CardCreate:
      type: object
      properties:
        label: { type: string, maxLength: 60, description: Shown in the app and statements. }
        form: { type: string, enum: [virtual], default: virtual, description: Physical cards land Phase 2. }
    Card:
      type: object
      required: [id, agent_id, form, status, last4, created_at]
      properties:
        id: { type: string, examples: [crd_v_01JABCDXYZ] }
        agent_id: { type: string }
        label: { type: [string, 'null'] }
        form: { type: string, enum: [virtual, physical] }
        status: { type: string, enum: [active, frozen, terminated] }
        network: { type: string, enum: [visa], default: visa }
        last4: { type: string }
        expiry_month: { type: integer }
        expiry_year: { type: integer }
        created_at: { type: string, format: date-time }
...